% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{aes_color}
\alias{aes_color}
\title{aes_color}
\usage{
aes_color(color, aes_som)
}
\arguments{
\item{color}{Boolean type to adding color on plot}

\item{aes_som}{Function aes_som for input data on plot}
}
\value{
Base plot
}
\description{
Function that return the base plot with or without color
}
