% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_som_result.R
\name{generate.model.result}
\alias{generate.model.result}
\title{generate.model.som}
\usage{
generate.model.result(model.som, cutree = FALSE, cutree.value = 0,
  grid.mark = FALSE)
}
\arguments{
\item{model.som}{Object of SOM package}

\item{cutree}{Boolean type to set of cluster visualization}

\item{cutree.value}{Pieces of division grid}

\item{grid.mark}{Boolean type to mark amount of neuron in each grid}
}
\value{
Model data that is used in plots
}
\description{
generate.model.som
}
