% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsom_entropy.R
\name{ggsom_entropy}
\alias{ggsom_entropy}
\title{Function to obtain the purity of each neuron in the SOM network}
\usage{
ggsom_entropy(ggsom_aes)
}
\arguments{
\item{ggsom_aes}{kohonen package object modeling}
}
\value{
Data set with the purity attribute added in Tibble
}
\description{
Entropy calculation using the maximum likelihood method
}
\author{
Felipe Carvalho, \email{felipe.carvalho@inpe.br}
}
