% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsom_line.R
\name{ggsom_line}
\alias{ggsom_line}
\title{plot_ggsom_line_point_colour}
\usage{
ggsom_line(aes_som, color)
}
\arguments{
\item{color}{Boolean to change color of the edges}

\item{ggsom}{Model of data plot}
}
\value{
ggplot visualization
}
\description{
Visualization that is contains amount of neuron per grid with color based on classes
}
