% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsom_cluster.R
\name{ggsom_rect}
\alias{ggsom_rect}
\title{Cluster plot using geom rect with color}
\usage{
ggsom_rect(ggsom, text = TRUE)
}
\arguments{
\item{ggsom}{Model of data plot}

\item{text}{Boolean type to viasualize numbers of neurons per grid}
}
\value{
ggplot visualization
}
\description{
Cluster plot using geom rect with color
}
