% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsom_ts.R
\name{ggsom_ts}
\alias{ggsom_ts}
\title{Time series base visualization on som packages}
\usage{
ggsom_ts(ggsom_aes, x_axis = days, y_axis = values,
  color = reference, x_map = "2017-09-02", y_map = 0.59)
}
\arguments{
\item{ggsom_aes}{Model generate by ggsom_aes function}

\item{x_axis}{x-axis of time series}

\item{y_axis}{y-axis of time series values}

\item{color}{the time series classes}

\item{x_map}{the position where the number of samples per class will be marked corresponds to the x axis}

\item{y_map}{the position at which the number of samples per class will be marked corresponds to the y axis}
}
\value{
ggplot2 class object
}
\description{
View user base for customizing SOM object time series charts
}
\references{
`ggplot2` package (https://CRAN.R-project.org/package=ggplot2)
}
\author{
Felipe Carvalho, \email{lipecaso@gmail.com}
}
