% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsom_utils.R
\name{is.kohonen}
\alias{is.kohonen}
\title{Verify if objects inherits kohonen object}
\usage{
is.kohonen(data)
}
\arguments{
\item{data}{Boolean type to generate plot of timeseries}
}
\value{
Boolean value
}
\description{
if object inherits kohonen class return TRUE otherwise stop
}
\references{
`Kohonen`package (https://CRAN.R-project.org/package=kohonen)
}
\author{
Felipe Carvalho, \email{lipecaso@gmail.com}
}
