% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsom_line.R
\name{plot_ggsom_line_point}
\alias{plot_ggsom_line_point}
\title{plot_ggsom_line_point}
\usage{
plot_ggsom_line_point(ggsom)
}
\arguments{
\item{ggsom}{Model of data plot}
}
\value{
ggplot visualization
}
\description{
Visualization that is contains amount of neuron per grid without colors
}
