% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsom_line.R
\name{plot_ggsom_line_point_colour}
\alias{plot_ggsom_line_point_colour}
\title{plot_ggsom_line_point_colour}
\usage{
plot_ggsom_line_point_colour(aes_som)
}
\arguments{
\item{ggsom}{Model of data plot}
}
\value{
ggplot visualization
}
\description{
Visualization that is contains amount of neuron per grid with color based on classes
}
