% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_som_result.R
\name{sum.unit.class}
\alias{sum.unit.class}
\title{Sum.unit.class}
\usage{
\method{sum}{unit.class}(model.som.values)
}
\arguments{
\item{model.som.values}{Model of values that is generate in generate.model.result}
}
\value{
Sum of amount neurons per grid
}
\description{
Sum of each grid the neuron
}
