% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aes_som.R
\name{sum_unit_class}
\alias{sum_unit_class}
\title{sum_unit_class}
\usage{
sum_unit_class(model_som_values)
}
\arguments{
\item{model_som_values}{Model of values that is generate in generate.model.result}
}
\value{
Sum of amount neurons per grid
}
\description{
Sum of each grid the neuron
}
