% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_data.R
\name{visualize.cluster.numbers}
\alias{visualize.cluster.numbers}
\title{visualize.cluster.numbers}
\usage{
visualize.cluster.numbers(model.som.result)
}
\arguments{
\item{model.som.result}{Model of data plot}
}
\value{
ggplot visualization
}
\description{
Visualization of each cluster using cutree and each amount of neuron per grid
}
