% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_data.R
\name{visualize.cluster}
\alias{visualize.cluster}
\title{Visualize.cluster}
\usage{
visualize.cluster(model.som.result, factor = FALSE)
}
\arguments{
\item{model.som.result}{Model of data plot}

\item{factor}{Boolean data type to set a different colours in plot}
}
\value{
ggplot visualization
}
\description{
Visualization of each cluster using cutree
}
