% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_data.R
\name{visualize.numbers}
\alias{visualize.numbers}
\title{visualize.numbers}
\usage{
visualize.numbers(model.som.result, colour = FALSE)
}
\arguments{
\item{model.som.result}{Model of data plot}

\item{colour}{Boolean data type to set colours in plot}
}
\value{
ggplot visualization
}
\description{
Visualization amount of neuron per grid
}
