% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aes_som.R
\name{aes_som}
\alias{aes_som}
\title{aesthetics data of som object}
\usage{
aes_som(model_som, cutree_value = 0)
}
\arguments{
\item{model_som}{object of SOM package}

\item{cutree.value}{number of division grid}
}
\value{
Model data that is used in plots
}
\description{
aesthetics data of som object
}
