% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsom_aes.R
\name{ggsom_aes}
\alias{ggsom_aes}
\title{kohonen package object modeling}
\usage{
ggsom_aes(object_som, class)
}
\arguments{
\item{object_som}{object of kohonen package}

\item{class}{categorical vector corresponding to the class of the dataset}
}
\value{
data.table model used in visualizations
}
\description{
Function to map each SOM neuron with its corresponding class
}
\references{
`Kohonen`package (https://CRAN.R-project.org/package=kohonen)
}
\author{
Felipe Carvalho, \email{lipecaso@gmail.com}
}
