% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsom_rect.R
\name{ggsom_rect}
\alias{ggsom_rect}
\title{Cluster plot using geom rect with color}
\usage{
ggsom_rect(aes_som, text = TRUE)
}
\arguments{
\item{aes_som}{Function aes_som for input data on plot}

\item{text}{Boolean type to viasualize numbers of neurons per grid}
}
\value{
ggplot visualization
}
\description{
Cluster plot using geom rect with color
}
\examples{

\dontrun{
library(ggthemes)
library(ggsom)
library(ggplot2)

iris_som <- som(scale(iris[1:4]), grid = somgrid(6, 4, "rectangular"))

ggsom_rect(aes_som(iris_som, cutree_value = 3), TRUE)

}

}
