% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsom_ribbon.R
\name{ggsom_ribbon}
\alias{ggsom_ribbon}
\title{Cluster plot using geom ribbon with color}
\usage{
ggsom_ribbon(ggsom, color)
}
\arguments{
\item{ggsom}{Model of data plot}

\item{color}{Boolean type to add color on plot}
}
\value{
ggplot visualization
}
\description{
Cluster plot using geom ribbon with color
}
