% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.cluster}
\alias{is.cluster}
\title{Check if the parameter has a cluster column}
\usage{
is.cluster(x)
}
\arguments{
\item{x}{Dataframe}
}
\value{
TRUE if cluster in data frame otherwise FALSE
}
\description{
Check if the parameter has a cluster column
}
