% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsom_utils.R
\name{is.kohonen}
\alias{is.kohonen}
\title{verifies that the object inherits kohonen object}
\usage{
is.kohonen(object_som)
}
\arguments{
\item{object_som}{object of Kohonen package}
}
\value{
Boolean value
}
\description{
if object inherits kohonen class return TRUE otherwise FALSE
}
\references{
`Kohonen`package (https://CRAN.R-project.org/package=kohonen)
}
\author{
Felipe Carvalho, \email{lipecaso@gmail.com}
}
