% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.kohonen}
\alias{is.kohonen}
\title{Check if the object is inherits of kohonen}
\usage{
is.kohonen(x)
}
\arguments{
\item{x}{Dataframe

This is used by aes_som to simplify comparison between objects}
}
\value{
if object inherits kohonen class return TRUE otherwise stop
}
\description{
Check if the object is inherits of kohonen
}
